WIP_Projects/advance_ai_agents/price_monitoring_agent/pyproject.toml
[project]
name = "price-monitoring-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.159.0",
    "litellm>=1.74.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scrapegraphai>=1.62.0",
    "sendgrid>=6.12.4",
    "streamlit>=1.48.1",
    "twilio>=9.7.0",
]
